Python program that can be used to calculate the area of a triangle version 2
Copy of the Area of a Triangle program created previously. Modify the program such that after receiving input from the user for the length of the base and the height of the triangle, it (the program) will verify that the values entered are not zero or negative. If they are, the program should output a useful message and then stop.