Investment Table Version 1
A program that will generate a compound interest table for an initial principal amount and a fixed annual interest  rate. The program must prompt the user to enter the amount initially invested, the annual rate and the number of years the principal is invested. Once these have been entered, the program will format a table showing the value of the investment and the end of each year.