Investment Table Version 2
A program that will determine the value of an investment where the interest rate varies year over year. 

The program will prompt the user of the principal amount invested. Then it will prompt the user of each year's interest rate. A sentinel value of an interest rate equal to 0 will be used to signal that the final year has been reached and the accumulated value of the investment is then output to the user. It will also display the average year income returned by the investment, calculated as the difference between the final balance and the principal invested, divided by the number of years of investment.